This commit might not be suitable for the TFT_eSPI library...
What I've implemented is a parallel interface for ES32 to a TFT display with ILI_9163 driver.

The reason for selecting the display is that it was the only one with
Sunligt Readable capability (1000cd), and it only has a parallel interface.

I would appreciate if the changes could somehow be incorporated into TFT_eSPI.

I have not implemented the routines that are reading from the display... as I do nto need them I think.

Please let me know what you think.

Regards,
Knut