For ESP32-S3 and TFT_PARALLEL_8_BIT, bitbang 20%-30% faster
So I was experimenting with my own library for this parallel setup
and my sole ESP32-S3 board and I was looking at the assembler output
for a single pixel write.  Before this hack, the best we could do
was 3 stores (s32i instructions) per byte written out to the TFT.

For each data byte, these register writes happen:

1. Pulls the WR line LOW in the same GPIO "clear" register write where
  one also sets the data lines which are low;

2. Perform a GPIO "set" register write for the high data lines.

3. Perform a third "set" register write just for pulling the WR line
high.

This commit, which for now is only a hack, conflates steps 2 and 3 for
any byte that is not the first byte after a command.

So only 2 stores per byte, which is theoretically a 33% maximum
improvement.

The stores in the Xtensa LX7 allow for some wait cycles (around 30 or
so), so it stills gives time to do the current run-time solution of
adjusting the result of the 'set_mask' lookup.  But with more lookup
tables even this could be avoided, so there's still room for
improvement.

* Processors/TFT_eSPI_ESP32_S3.c (pushBlock)
(pushSwapBytePixels)
(pushPixels): Use HACKIT(bool) macro.

* Processors/TFT_eSPI_ESP32_S3.h
(_hackit_): new global.
(SUPER_SPECIFIC_ESP32_S3_9BITPARALLEL_HACK, HACKIT): New defines
(tft_Write_8 (within define))
(tft_Write_16 (within define))
(tft_Write_16S (within define)): Use _hackit_.