Allowing -1 as a value for CONFIG_TFT_CS.
Context:
When the library get configured using the `idf.py menuconfig` tool(the
lib used as an ESP-IDF component), the CONFIG_TFT_CS config used for
selecting the pin used as the TFT_CS pin. However a check exists in the
`TFT_config.h` that raises an error if the value of CONFIG_TFT_CS is -1,
this check should be removed since -1 is needed as an value to allow manual
control for the CS pin (manual control for the CS pins is needed for
controlling more than one screen).